win32 implementation is now forced to use vsync to avoid burning cpu

+++ src/core/core.h
@@ -10,7 +10,7 @@ typedef struct _window_store window_store;
 typedef struct _global_store
 {
     int alive;
-    double gl_version;
+    int gl_version;
 
     window_store* window;
 }
